+++ src/main/java/com/leonyk2/mcmod/item/ModItems.java
@@ -22,7 +22,9 @@ public class ModItems {
             () -> new Item(new Item.Properties()));
 
     public static RegistryObject<Item> UNCOMPLETE_DIAMOND = ITEMS.register("incomplete_diamond",
-            () -> new Item(new Item.Properties()));
+            () -> new Item(new Item.Properties().stacksTo(1)));
+    public static RegistryObject<Item> UNCOMPLETE_DIAMOND_BLOCK = ITEMS.register("incomplete_diamond_block",
+            () -> new Item(new Item.Properties().stacksTo(1)));
 
     public static RegistryObject<Item> LINSENEINTOPF = ITEMS.register("linseneintopf",
             () -> new Item(new Item.Properties().food(ModFoods.LINSENEINTOPF)));